+++ src/ops/patterns.h
@@ -7,18 +7,29 @@ extern const tele_op_t op_P_N;
 extern const tele_op_t op_P;
 extern const tele_op_t op_PN;
 extern const tele_op_t op_P_L;
+extern const tele_op_t op_PN_L;
 extern const tele_op_t op_P_WRAP;
+extern const tele_op_t op_PN_WRAP;
 extern const tele_op_t op_P_START;
+extern const tele_op_t op_PN_START;
 extern const tele_op_t op_P_END;
+extern const tele_op_t op_PN_END;
 extern const tele_op_t op_P_I;
+extern const tele_op_t op_PN_I;
 extern const tele_op_t op_P_HERE;
+extern const tele_op_t op_PN_HERE;
 extern const tele_op_t op_P_NEXT;
+extern const tele_op_t op_PN_NEXT;
 extern const tele_op_t op_P_PREV;
+extern const tele_op_t op_PN_PREV;
 
 extern const tele_op_t op_P_INS;
+extern const tele_op_t op_PN_INS;
 extern const tele_op_t op_P_RM;
+extern const tele_op_t op_PN_RM;
 extern const tele_op_t op_P_PUSH;
+extern const tele_op_t op_PN_PUSH;
 extern const tele_op_t op_P_POP;
-
+extern const tele_op_t op_PN_POP;
 
 #endif